+++ src/Arango/Arango.Client/Protocol/Request.cs
@@ -12,14 +12,16 @@ namespace Arango.Client.Protocol
         internal Dictionary<string, string> QueryString = new Dictionary<string, string>();
         internal string Body { get; set; }
 
-        internal Request(HttpMethod httpMethod, string apiUri)
+        internal Request(HttpMethod httpMethod, string apiUri) : this(httpMethod, apiUri, "")
+        {
+        }
+
+        internal Request(HttpMethod httpMethod, string apiUri, string operationUri)
         {
             HttpMethod = httpMethod;
             OperationUri = apiUri;
         }
 
-        internal Request(HttpMethod httpMethod, string apiUri, string operationUri) : this(httpMethod, apiUri + operationUri) {}
-        
         internal string GetRelativeUri()
         {
             var uri = new StringBuilder(OperationUri);