+++ src/Arango/Arango.ConsoleTests/Program.cs
@@ -1,7 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
-using Arango.Client;
-using Arango.fastJSON;
 
 namespace Arango.ConsoleTests
 {
@@ -9,67 +6,13 @@ namespace Arango.ConsoleTests
     {
         public static void Main(string[] args)
         {
-            /*var dummy = new Dummy
-            {
-                Foo = "whoa",
-                Bar = 123,
-                Str = "small str",
-                Ignore = "should be igored",
-                IgnoreNullValue = null
-            };
-
-            var converted = DictionaryExtensions.StripObject(dummy);
-
-            Console.WriteLine(JSON.ToJSON(converted, new JSONParameters { UseEscapedUnicode = false, UseFastGuid = false, UseExtensions = false }));*/
-
-            //InsertTest();
-            PerformanceTests();
-            
-            /*JsonToObjectHeatUp();
-            JsonToObjectAsDictionaryHeatup();
-            JsonParseHeatUp();
-            
-            for (int i = 0; i < 100; i++)
-            {
-                Console.WriteLine("Iteration: {0}", i + 1);
-                
-                JsonToObjectTest();
-                JsonToObjectAsDictionaryTest();
-                JsonParseTest();
-            }*/
+            //PerformanceTests();
+            FoxxPerformanceTests();
             
             Console.Write("Press any key to continue . . . ");
             Console.ReadKey(true);
         }
-        
-        static void InsertTest()
-        {
-            Database.CleanupTestDatabases();
-            Database.CreateTestDatabase(Database.TestDatabaseGeneral);
-            Database.CreateTestCollection("Release", ACollectionType.Document);
-            
-            var db = new ADatabase(Database.Alias);
-            
-            var album = new Release { Id = 123 , Status = "some status", Title = "some title", Country = "some country", Released = "released" };
-            var releaseKey = CreateDocument(db, "Release", album);
-        }
-        
-        static string CreateDocument(ADatabase db, string collection, object dataType)
-        {
-            var createDocumentResult = db.Document
-                .WaitForSync(false)
-                .Create(collection, dataType);
-            
-            var key = "";
-            
-            if (createDocumentResult.Success)
-            {
-                key = createDocumentResult.Value.String("_key");
-            }
-            
-            return key;
-        }
-        
+
         static void PerformanceTests()
         {
             var iterationCount = 10000;
@@ -83,106 +26,13 @@ namespace Arango.ConsoleTests
 
             performance.Dispose();
         }
-        
-        static void JsonToObjectHeatUp()
-        {
-            var json = "{\"Foo\":\"some string\",\"Bar\":123}";
-            
-            //var obj = JSON.ToObject<Dictionary<string, object>>(json);
-            var obj = JSON.ToObject<Dummy>(json);
-            
-            Console.WriteLine(obj);
-        }
-        
-        static void JsonToObjectTest()
-        {
-            var json = "{\"Foo\":\"some string\",\"Bar\":123}";
-            
-            //var obj = JSON.ToObject<Dictionary<string, object>>(json);
-            var obj = JSON.ToObject<Dummy>(json);
-            
-            Console.WriteLine(obj);
-        }
-        
-        static void JsonToObjectAsDictionaryHeatup()
-        {
-            var json = "{\"Foo\":\"some string\",\"Bar\":123}";
-            
-            //var obj = JSON.ToObject<Dictionary<string, object>>(json);
-            var obj = JSON.ToObject<Dictionary<string, object>>(json);
-            
-            Console.WriteLine(obj);
-        }
-        
-        static void JsonToObjectAsDictionaryTest()
-        {
-            var json = "{\"Foo\":\"some string\",\"Bar\":123}";
-            
-            //var obj = JSON.ToObject<Dictionary<string, object>>(json);
-            var obj = JSON.ToObject<Dictionary<string, object>>(json);
-            
-            Console.WriteLine(obj);
-        }
-        
-        static void JsonParseHeatUp()
-        {
-            var json = "{\"Foo\":\"some string\",\"Bar\":123}";
-            
-            //var obj = JSON.Parse(json) as Dictionary<string, object>;
-            var obj = GetObject<Dummy>(JSON.Parse(json) as Dictionary<string, object>);
-            
-            Console.WriteLine(obj);
-        }
-        
-        static void JsonParseTest()
-        {
-            var json = "{\"Foo\":\"some string\",\"Bar\":123}";
-            
-            //var obj = JSON.Parse(json) as Dictionary<string, object>;
-            var obj = GetObject<Dummy>(JSON.Parse(json) as Dictionary<string, object>);
-            
-            Console.WriteLine(obj);
-        }
-        
-        static Object GetObject(Dictionary<string, object> dict, Type type)
-        {
-            var obj = Activator.CreateInstance(type);
 
-            foreach (var kv in dict)
-            {
-                var prop = type.GetProperty(kv.Key);
-                if(prop == null) continue;
-    
-                object value = kv.Value;
-                if (value is Dictionary<string, object>)
-                {
-                    value = GetObject((Dictionary<string, object>) value, prop.PropertyType); // <= This line
-                }
-    
-                prop.SetValue(obj, value, null);
-            }
-            
-            return obj;
-        }
-        
-        static T GetObject<T>(Dictionary<string, object> dict)
+        static void FoxxPerformanceTests()
         {
-            return (T)GetObject(dict, typeof(T));
+            var iterationCount = 10000;
+            var performance = new FoxxPerformance();
+
+            performance.TestPostCall(iterationCount);
         }
-    }
-    
-    public class Dummy
-    {
-        public string Foo { get; set; }
-        public long Bar { get; set; }
-
-        [AliasField("str")]
-        public string Str { get; set; }
-
-        [IgnoreField]
-        public string Ignore { get; set; }
-
-        [IgnoreNullValue]
-        public object IgnoreNullValue { get; set; }
     }
 }