+++ src/Arango/Arango.Tests/Entities/TransactionEntity.cs
@@ -3,6 +3,7 @@ namespace Arango.Tests
 {
     public class TransactionEntity
     {
+        public string Foo { get; set; }
         public bool Executed { get; set; }
     }
 }